Former-commit-id: 77aa57250ae7eba95e5b0268a092daa7926889ea

+++ libraries/classification/training_set.h
@@ -205,7 +205,7 @@ public:
     /**
      * Returns the number of items in the training set
      */
-    unsigned size() const
+    size_t size() const
     {
         return theItems.size();
     };
